[REF] Add Loggable, Happenable and Timelineable Concerns [IMP] Update timeline views to enable other kind of happenings, besides BlogPosts [IMP] Use nested_attributes for happenings

+++ app/helpers/menu_helper.rb
@@ -23,27 +23,32 @@ module MenuHelper
   def crud_menu_options(resource, additional_items = [])
     link_items = [].concat(additional_items).compact
     resource_policy = policy(resource)
+    if policy(BlogPost).create? && (resource.is_a?(Motion) || resource.is_a?(Question))
+      link_items << link_item(t('blog_posts.type_new'), polymorphic_url([:new, resource, :blog_post]), fa: 'quote-left')
+    end
     link_items << link_item(t('edit'), polymorphic_url([:edit, resource]), fa: 'edit') if resource_policy.update?
     link_items << link_item(t('log'), log_url(resource.edge), fa: 'history') if resource_policy.log?
-    if resource.is_trashed?
-      if resource_policy.trash?
-        link_items << link_item(t('untrash'),
-                                polymorphic_url([:untrash, resource]),
-                                data: {confirm: t('untrash_confirmation'), method: 'put', turbolinks: 'false'},
-                                fa: 'eye')
-      end
-      if resource_policy.destroy?
-        link_items << link_item(t('destroy'),
-                                polymorphic_url(resource, destroy: true),
-                                data: {confirm: t('destroy_confirmation'), method: 'delete', turbolinks: 'false'},
-                                fa: 'close')
-      end
-    else
-      if resource_policy.trash?
-        link_items << link_item(t('trash'),
-                                polymorphic_url(resource),
-                                data: {confirm: t('trash_confirmation'), method: 'delete', turbolinks: 'false'},
-                                fa: 'trash')
+    if resource.is_trashable?
+      if resource.is_trashed?
+        if resource_policy.trash?
+          link_items << link_item(t('untrash'),
+                                  polymorphic_url([:untrash, resource]),
+                                  data: {confirm: t('untrash_confirmation'), method: 'put', turbolinks: 'false'},
+                                  fa: 'eye')
+        end
+        if resource_policy.destroy?
+          link_items << link_item(t('destroy'),
+                                  polymorphic_url(resource, destroy: true),
+                                  data: {confirm: t('destroy_confirmation'), method: 'delete', turbolinks: 'false'},
+                                  fa: 'close')
+        end
+      else
+        if resource_policy.trash?
+          link_items << link_item(t('trash'),
+                                  polymorphic_url(resource),
+                                  data: {confirm: t('trash_confirmation'), method: 'delete', turbolinks: 'false'},
+                                  fa: 'trash')
+        end
       end
     end
     dropdown_options(t('menu'), [{items: link_items}], fa: 'fa-ellipsis-v')
